refactor(meili): üks jagatud build_work_documents mõlemale indekseerimisteele (#23)
Varem ehitasid live-tee (sync_work_to_meilisearch) ja seed-tee (1-1_consolidate)
dokumendid kahe paralleelse work_ctx + lehe-tsükliga, mis kutsusid sama
_build_page_document-i. Nüüd on KOGU teose-dokumendi ehitamine ühes funktsioonis.

- server/meili_doc.py: lisatud get_work_metadata (AINULT V2 — V1 fallback eemaldatud,
  vt allpool) + build_work_documents (metaandmed + work_ctx + lehe-tsükkel).
- meilisearch_ops.sync_work_to_meilisearch: kutsub build_work_documents'i, jätab
  alles ainult live-spetsiifilise saba (konfig + _upsert_work_documents Meili API-le).
  ~150 rida inline-loogikat eemaldatud.
- 1-1_consolidate_data.py: get_work_metadata/build_work_documents/sanitize_id
  eemaldatud (imporditud meili_doc-ist); jääb ainult JSONL-saba.

V1-fallback eemaldatud: server-skann (issue #23) kinnitas 0 V1-teost serveris
(pealkiri/aasta/koht/trükkal/teose_tags/autor — kõik 0; teose_id 2 teoses, ei
mõjuta väljundit). Mõlemad teed loevad nüüd puhast V2-d.

Väljund-neutraalsus verifitseeritud serveri päris andmetel: uue koodi ehitatud
21814 dok vs praegune indeks → 0 SISULIST erinevust (ainult alias-massiivide
JÄRJEKORD erineb, mis on otsingule ükskõik ja tuleneb person_aliases.json
taustauuendusest, mitte sellest refaktorist). Seetõttu reseed POLE vajalik.

Kogu suite: 525 passed (snapshot lukustab live-väljundi väli-väljalt).
